+++ app/src/main/java/com/example/weather/weather_model/Temp.kt
@@ -1,7 +1,7 @@
 package com.example.weather.weather_model
 
 data class Temp(val min: Double = 0.0,
-                val max: Int = 0,
+                val max: Double = 0.0,
                 val eve: Double = 0.0,
                 val night: Double = 0.0,
                 val day: Double = 0.0,
